refactor: styles landing
