+++ frontend/src/App.tsx
@@ -4,7 +4,10 @@ import { Home } from "./pages/Home"
 
 function App() {
   return (
-    <><Header /><Home /></>
+    <>
+      <Header />
+      <Home />
+    </>
   )
 }
